+++ packages/react-broadcast/modules/hocs/branch-on-feature-toggle/branch-on-feature-toggle.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { renderWithAdapter, components } from '@flopflip/test-utils';
 import branchOnFeatureToggle from './branch-on-feature-toggle';
-import Configure from '../configure';
+import Configure from '../../components/configure';
 
 const render = (TestComponent) =>
   renderWithAdapter(TestComponent, {
